Cleaned up the kingspray brush, get uv's working in the correct direction (so it's no longer upsidedown) and commented the code a bit.
Also added in a node that samples from the light attenuation buffer.  I
was able to make shadows affect the brush BUT it only used shadows from
the non-moveable light in the scene... not from the little spotlight I
tacked on to the top of the can.

I assume this is because a dynamic/moveable light can't cast shadows from
a moveable object?

There may be a solution using a fixed directional light for each paint
surface but that'll likely not work on 3d objects so probably not worth
trying.